[VP] fix_coverity_2
fix coverity

+++ media_softlet/agnostic/common/vp/hal/packet/vp_render_hdr_kernel.cpp
@@ -4156,7 +4156,7 @@ MOS_STATUS VpRenderHdrKernel::GetCurbeState(void *&curbe, uint32_t &curbeLength)
         m_hdrCurbe.DW58.TwoLayerOperationLayer0          = VPHAL_HDR_TWO_LAYER_OPTION_COMP;
     }
 
-    m_hdrCurbe.DW63.FormatDescriptorDestination        = FormatDescriptor;
+    m_hdrCurbe.DW63.FormatDescriptorDestination        = (uint32_t)FormatDescriptor;
     m_hdrCurbe.DW63.ChromaSittingLocationDestination   = ChromaSiting;
     m_hdrCurbe.DW63.ChannelSwapEnablingFlagDestination = bChannelSwap;
